refactor: extract YouTube URL capture strategy
