+++ src/background/post-url-capture.test.ts
@@ -1,9 +1,7 @@
 import { describe, expect, it } from 'vitest';
-import { Window } from 'happy-dom';
 import {
   buildPostUrlCaptureRetryPlan,
   buildPostUrlCaptureScriptArgs,
-  captureYouTubeStudioPostUrlInPage,
 } from './post-url-capture';
 
 describe('post URL capture retry plan', () => {
@@ -77,33 +75,5 @@ describe('post URL capture script arguments', () => {
       undefined,
       Number.NaN,
     )).toEqual(['youtube', 'short title', null, null]);
-  });
-});
-
-describe('YouTube Studio post URL capture', () => {
-  it('finds the matching video card without depending on the localized dashboard heading', async () => {
-    const window = new Window();
-    window.document.body.innerHTML = `
-      <a href="https://studio.youtube.com/video/older-video/edit">Older video</a>
-      <section>
-        <div>Latest video performance</div>
-        <article>
-          <h2 id="title">tutti surface matrix video 2026-07-25T23-42-46</h2>
-          <div>
-            <a href="https://studio.youtube.com/video/new-video-id/analytics/tab-overview">
-              Analytics
-            </a>
-          </div>
-        </article>
-      </section>
-    `;
-
-    await expect(captureYouTubeStudioPostUrlInPage(
-      'tutti surface matrix video 2026-07-25',
-      window.document as unknown as ParentNode,
-    )).resolves.toEqual({
-      url: 'https://www.youtube.com/watch?v=new-video-id',
-      trace: ['matched target title in scoped video card (attempt=0, depth=1)'],
-    });
   });
 });
